[ISSUE-87] Refactored material eval into separate struct
