+++ brain/src/pos.rs
@@ -1,6 +1,4 @@
-use crate::tables::PositionTables;
-use crate::values::PieceValues;
-use crate::{EvalBoardImpl, EvalParameters};
+use crate::{EvalBoardImpl, EvalParameters, PieceValues, PositionTables};
 use myopic_board::MutBoardImpl;
 
 /// Construct an instance of the default EvalBoard implementation using the
@@ -10,7 +8,7 @@ pub fn from_fen_and_params(
     params: EvalParameters,
 ) -> Result<EvalBoardImpl<MutBoardImpl>, String> {
     myopic_board::fen_position(fen)
-        .map(|inner| EvalBoardImpl::new(inner, params.position_tables, params.piece_values))
+        .map(|inner| EvalBoardImpl::new(inner, params.piece_values, params.position_tables))
 }
 
 /// Construct an instance of the default EvalBoard implementation using the
@@ -24,7 +22,7 @@ pub fn from_pgn_and_params(
     params: EvalParameters,
 ) -> Result<EvalBoardImpl<MutBoardImpl>, String> {
     myopic_board::parse::position_from_pgn(pgn_sequence)
-        .map(|inner| EvalBoardImpl::new(inner, params.position_tables, params.piece_values))
+        .map(|inner| EvalBoardImpl::new(inner, params.piece_values, params.position_tables))
 }
 
 pub fn from_pgn(pgn_sequence: &str) -> Result<EvalBoardImpl<MutBoardImpl>, String> {
@@ -36,7 +34,7 @@ pub fn from_uci_and_params(
     params: EvalParameters,
 ) -> Result<EvalBoardImpl<MutBoardImpl>, String> {
     myopic_board::parse::position_from_uci(uci_sequence)
-        .map(|inner| EvalBoardImpl::new(inner, params.position_tables, params.piece_values))
+        .map(|inner| EvalBoardImpl::new(inner, params.piece_values, params.position_tables))
 }
 
 pub fn from_uci(uci_sequence: &str) -> Result<EvalBoardImpl<MutBoardImpl>, String> {
@@ -46,7 +44,7 @@ pub fn from_uci(uci_sequence: &str) -> Result<EvalBoardImpl<MutBoardImpl>, Strin
 pub fn start() -> EvalBoardImpl<MutBoardImpl> {
     EvalBoardImpl::new(
         myopic_board::start_position(),
-        PositionTables::default(),
         PieceValues::default(),
+        PositionTables::default(),
     )
 }
